video: fix i-frame-interval can not set to float.
STEPS TO REPRODUCE: (please be specific)
1. APP start a video record , and set the i-frame-interval to float like 0.033, set the frame rate to 240fps
2. Analysis if the i-frame-interval is ok
while the framerate is 240fps, and the i-frame-interval is set to 0.033s.
the EXPECTED RESULTS: there is 1 i-frame in 8 frames.
but the OBSERVED RESULTS: there is 1 i-frame in 240 frames

Bug: 183576655

+++ media/libmediaplayerservice/StagefrightRecorder.h
@@ -125,7 +125,7 @@ private:
     int32_t mAudioChannels;
     int32_t mSampleRate;
     int32_t mInterleaveDurationUs;
-    int32_t mIFramesIntervalSec;
+    float   mIFramesIntervalSec;
     int32_t mCameraId;
     int32_t mVideoEncoderProfile;
     int32_t mVideoEncoderLevel;
@@ -220,7 +220,7 @@ private:
     status_t setParamCaptureFps(double fps);
     status_t setParamVideoEncodingBitRate(int32_t bitRate);
     status_t setParamVideoBitRateMode(int32_t bitRateMode);
-    status_t setParamVideoIFramesInterval(int32_t seconds);
+    status_t setParamVideoIFramesInterval(float seconds);
     status_t setParamVideoEncoderProfile(int32_t profile);
     status_t setParamVideoEncoderLevel(int32_t level);
     status_t setParamVideoCameraId(int32_t cameraId);
